requirements.txt
openai
pillow
numpy
jsonschema